+++ src/js/defaults.js
@@ -23,6 +23,8 @@ const pluginDefaultOptions = {
     // Boolean indicating if milliseconds should be included,
     // e.g. "00:00:000" vs "00:00".
     displayMilliseconds: false,
+    // Use a custom time format function, for example: (seconds, guide) => `test:${seconds}:${guide}`
+    formatTime: undefined,
     // Width of the recorded video frames.
     frameWidth: 320,
     // Height of the recorded video frames.
@@ -55,7 +57,8 @@ const pluginDefaultOptions = {
     // disable (default).
     videoWebAssemblyURL: '',
     // Audio recording library to use. Legal values are 'recordrtc' (default),
-    // 'libvorbis.js', 'opus-recorder', 'lamejs', 'vmsg' and 'recorder.js'.
+    // 'libvorbis.js', 'opus-recorder', 'opus-media-recorder', 'lamejs', 'vmsg'
+    // and 'recorder.js'.
     audioEngine: 'recordrtc',
     // Audio recorder type to use. This allows you to specify an alternative
     // recorder class, e.g. StereoAudioRecorder. Defaults to 'auto' which let's
@@ -122,8 +125,8 @@ const pluginDefaultOptions = {
     imageOutputQuality: 0.92,
     // Accepts numbers in milliseconds; use this to force intervals-based blobs.
     timeSlice: 0,
-    // Media converter library to use. Legal values are 'ts-ebml' and 'ffmpeg.js'.
-    // Use an empty string '' to disable (default).
+    // Media converter library to use. Legal values are 'ts-ebml', 'ffmpeg.wasm'
+    // and 'ffmpeg.js'. Use an empty string '' to disable (default).
     convertEngine: '',
     // URL for the converter worker.
     convertWorkerURL: '',
@@ -132,7 +135,7 @@ const pluginDefaultOptions = {
     // Enable keyboard hotkeys.
     hotKeys: false,
     // Use this object to specify additional settings for the library used by the
-    // plugin (only used in opus-recorder and vmsg plugins).
+    // plugin (only used in opus-recorder, ffmpeg.js, ffmpeg.wasm and vmsg plugins).
     pluginLibraryOptions: {}
 };