+++ src/js/engine/engine-loader.js
@@ -6,8 +6,8 @@
 import videojs from 'video.js';
 
 import RecordRTCEngine from './record-rtc';
-import {CONVERT_PLUGINS, TSEBML, FFMPEGJS} from './convert-engine';
-import {RECORDRTC, LIBVORBISJS, RECORDERJS, LAMEJS, OPUSRECORDER, VMSG, WEBMWASM, GIFSHOT, AUDIO_PLUGINS} from './record-engine';
+import {CONVERT_PLUGINS, TSEBML, FFMPEGJS, FFMPEGWASM} from './convert-engine';
+import {RECORDRTC, LIBVORBISJS, RECORDERJS, LAMEJS, OPUSRECORDER, OPUSMEDIARECORDER, VMSG, WEBMWASM, GIFSHOT, AUDIO_PLUGINS} from './record-engine';
 
 /**
  * Get audio plugin engine class.
@@ -42,6 +42,11 @@ const getAudioEngine = function(audioEngine) {
         case OPUSRECORDER:
             // opus-recorder
             AudioEngineClass = videojs.OpusRecorderEngine;
+            break;
+
+        case OPUSMEDIARECORDER:
+            // opus-media-recorder
+            AudioEngineClass = videojs.OpusMediaRecorderEngine;
             break;
 
         case VMSG:
@@ -128,6 +133,11 @@ const getConvertEngine = function(convertEngine) {
         case FFMPEGJS:
             // ffmpeg.js
             ConvertEngineClass = videojs.FFmpegjsEngine;
+            break;
+
+        case FFMPEGWASM:
+            // ffmpeg.wasm
+            ConvertEngineClass = videojs.FFmpegWasmEngine;
             break;
 
         default: