+++ src/js/videojs.record.js
@@ -66,14 +66,19 @@ class Record extends Plugin {
         // (re)set recorder state
         this.resetState();
 
-        // use custom video.js time format
-        videojs.setFormatTime((seconds, guide) => {
-            return formatTime(seconds, guide, this.displayMilliseconds);
-        });
+        // use custom time format for video.js player
+        if (options.formatTime && typeof options.formatTime === 'function') {
+            // user-supplied formatTime
+            this.setFormatTime(options.formatTime);
+        } else {
+            // plugin's default formatTime
+            this.setFormatTime((seconds, guide) => {
+                return formatTime(seconds, guide, this.displayMilliseconds);
+            });
+        }
 
         // add device button with icon based on type
         let deviceIcon = 'av-perm';
-
         switch (this.getRecordType()) {
             case IMAGE_ONLY:
             case VIDEO_ONLY:
@@ -263,6 +268,9 @@ class Record extends Plugin {
             case AUDIO_ONLY:
                 // reference to videojs-wavesurfer plugin
                 this.surfer = this.player.wavesurfer();
+
+                // use same time format as this plugin
+                this.surfer.setFormatTime(this._formatTime);
                 break;
 
             case IMAGE_ONLY:
@@ -1224,7 +1232,7 @@ class Record extends Plugin {
                     // update current time display component
                     this.player.controlBar.currentTimeDisplay.formattedTime_ =
                         this.player.controlBar.currentTimeDisplay.contentEl().lastChild.textContent =
-                            formatTime(this.streamCurrentTime, duration, this.displayMilliseconds);
+                            this._formatTime(this.streamCurrentTime, duration, this.displayMilliseconds);
                 }
                 break;
         }
@@ -1268,7 +1276,7 @@ class Record extends Plugin {
                     this.player.controlBar.durationDisplay.contentEl().lastChild) {
                     this.player.controlBar.durationDisplay.formattedTime_ =
                     this.player.controlBar.durationDisplay.contentEl().lastChild.textContent =
-                        formatTime(duration, duration, this.displayMilliseconds);
+                        this._formatTime(duration, duration, this.displayMilliseconds);
                 }
                 break;
         }
@@ -1307,20 +1315,32 @@ class Record extends Plugin {
     }
 
     /**
-     * Show save as dialog in browser so the user can store the recorded media
-     * locally.
+     * Show save as dialog in browser so the user can store the recorded or
+     * converted media locally.
      *
-     * @param {object} name - Object with one or more names for the particular
-     *     blob(s) you want to save. File extensions are added automatically.
-     *     For example: {'video': 'name-of-video-file'}. Supported keys are
+     * @param {Object} name - Object with names for the particular blob(s)
+     *     you want to save. File extensions are added automatically. For
+     *     example: {'video': 'name-of-video-file'}. Supported keys are
      *     'audio', 'video' and 'gif'.
+     * @param {String} type - Type of media to save. Legal values are 'record'
+     *     (default) and 'convert'.
      * @example
-     * // save video file as 'foo.webm'
+     * // save recorded video file as 'foo.webm'
      * player.record().saveAs({'video': 'foo'});
+     *
+     * // save converted video file as 'bar.mp4'
+     * player.record().saveAs({'video': 'bar'}, 'convert');
+     * @returns {void}
      */
-    saveAs(name) {
-        if (this.engine && name !== undefined) {
-            this.engine.saveAs(name);
+    saveAs(name, type = 'record') {
+        if (type === 'record') {
+            if (this.engine && name !== undefined) {
+                this.engine.saveAs(name);
+            }
+        } else if (type === 'convert') {
+            if (this.converter && name !== undefined) {
+                this.converter.saveAs(name);
+            }
         }
     }
 
@@ -1821,6 +1841,25 @@ class Record extends Plugin {
 
         // error if we get here: notify listeners
         this.player.trigger(Event.ERROR, errorMessage);
+    }
+
+    /**
+     * Replaces the default `formatTime` implementation with a custom implementation.
+     *
+     * @param {function} customImplementation - A function which will be used in place
+     *     of the default `formatTime` implementation. Will receive the current time
+     *     in seconds and the guide (in seconds) as arguments.
+     */
+    setFormatTime(customImplementation) {
+        this._formatTime = customImplementation;
+
+        videojs.setFormatTime(this._formatTime);
+
+        // audio-only
+        if (this.surfer) {
+            // use same time format as this plugin
+            this.surfer.setFormatTime(this._formatTime);
+        }
     }
 
     /**